chore: rewrite SevenTV
