+++ layers/emote-integrations/layers/integrations/seventv/server/utils/types.ts
@@ -12,57 +12,13 @@ export interface __Emote extends IEmote {
   url: `https://cdn.7tv.app/emote/${string}/1x.webp`;
 }
 
-// eslint-disable-next-line @typescript-eslint/no-namespace
-export namespace TSevenTV {
-  export interface GlobalEmote extends __Emote {
-    type: "global";
-  }
-  export interface ChannelEmote extends __Emote {
-    type: "channel";
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  export namespace Global {
-    export type Emote = GlobalEmote;
-
-    export type Set = IEmoteSet<Emote>;
-
-    export interface ReadyIntegration extends TEmoteIntegrations.Global.Ready {
-      sets: Set[];
-      source: "SevenTV";
-    }
-
-    export interface FailedIntegration
-      extends TEmoteIntegrations.Global.Failed {
-      source: "SevenTV";
-    }
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  export namespace Person {
-    export interface ChannelSet extends IEmoteSet {
-      emotes: ChannelEmote[];
-    }
-
-    export type Set = ChannelSet;
-
-    export interface IntegrationOwner extends IEmoteIntegrationOwner {
-      avatarUrl: string;
-      id: string;
-      twitch: PersonTwitch;
-      pageAddress: `https://betterttv.com/users/${string}`;
-    }
-
-    export interface ReadyIntegration extends TEmoteIntegrations.Person.Ready {
-      sets: Set[];
-      owner: IntegrationOwner;
-      source: "SevenTV";
-    }
-
-    export interface FailedIntegration
-      extends TEmoteIntegrations.Person.Failed {}
-
-    export type Integration = ReadyIntegration | FailedIntegration;
-    export type SettledIntegration = ReadyIntegration | FailedIntegration;
-  }
+export interface GlobalEmote extends __Emote {
+  type: "global";
 }
+
+export interface ChannelEmote extends __Emote {
+  type: "channel";
+}
+
+export * as Person from "./types.person";
+export * as Global from "./types.global";
